+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to
 
 ### Fixed
 
+- Fix history page filters not working for rejected work orders
+  [#2391](https://github.com/OpenFn/lightning/issues/2391)
+
 ## [v2.9.4] - 2024-09-16
 
 ### Changed